requirements.txt
python-telegram-bot==20.7
psutil==5.9.7
Pillow==10.1.0
keyboard==0.13.5
pyautogui==0.9.54
pywin32==306
httpx==0.25.2
nest-asyncio==1.5.8
python-dotenv==1.0.0
aiohttp==3.9.1
cryptography==41.0.7
pynput==1.7.6
mss==9.0.1
opencv-python-headless==4.8.1.78
numpy==1.26.2
requests==2.31.0
packaging==23.2
dearpygui==1.10.1
typing-extensions==4.9.0
pyarmor==9.0.6
pyinstaller==6.3.0
wmi==1.5.1
